requirements.txt
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.20
langchain-google-genai==0.0.8
faiss-cpu==1.12.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
transformers==4.31.0
tokenizers==0.13.3
huggingface_hub==0.14.1
nltk==3.8.1
pypdf>=3.0.0
redis>=4.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# .\.venv311\Scripts\Activate.ps1
